+++ src/Symfony/Component/Messenger/Tests/WorkerTest.php
@@ -22,11 +22,14 @@ use Symfony\Component\Messenger\Retry\RetryStrategyInterface;
 use Symfony\Component\Messenger\Stamp\ReceivedStamp;
 use Symfony\Component\Messenger\Stamp\RedeliveryStamp;
 use Symfony\Component\Messenger\Stamp\SentStamp;
-use Symfony\Component\Messenger\Tests\Fixtures\CallbackReceiver;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
+use Symfony\Component\Messenger\Transport\Receiver\ReceiverInterface;
 use Symfony\Component\Messenger\Worker;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
+/**
+ * @group time-sensitive
+ */
 class WorkerTest extends TestCase
 {
     public function testWorkerDispatchTheReceivedMessage()
@@ -34,18 +37,22 @@ class WorkerTest extends TestCase
         $apiMessage = new DummyMessage('API');
         $ipaMessage = new DummyMessage('IPA');
 
-        $receiver = new CallbackReceiver(function ($handler) use ($apiMessage, $ipaMessage) {
-            $handler(new Envelope($apiMessage));
-            $handler(new Envelope($ipaMessage));
-        });
+        $receiver = new DummyReceiver([
+            [new Envelope($apiMessage), new Envelope($ipaMessage)],
+        ]);
 
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
 
         $bus->expects($this->at(0))->method('dispatch')->with($envelope = new Envelope($apiMessage, new ReceivedStamp()))->willReturn($envelope);
         $bus->expects($this->at(1))->method('dispatch')->with($envelope = new Envelope($ipaMessage, new ReceivedStamp()))->willReturn($envelope);
 
-        $worker = new Worker($receiver, $bus, 'receiver_id');
-        $worker->run();
+        $worker = new Worker([$receiver], $bus);
+        $worker->run([], function (?Envelope $envelope) use ($worker) {
+            // stop after the messages finish
+            if (null === $envelope) {
+                $worker->stop();
+            }
+        });
 
         $this->assertSame(2, $receiver->getAcknowledgeCount());
     }
@@ -53,24 +60,26 @@ class WorkerTest extends TestCase
     public function testWorkerDoesNotWrapMessagesAlreadyWrappedWithReceivedMessage()
     {
         $envelope = new Envelope(new DummyMessage('API'));
-        $receiver = new CallbackReceiver(function ($handler) use ($envelope) {
-            $handler($envelope);
-        });
+        $receiver = new DummyReceiver([[$envelope]]);
         $envelope = $envelope->with(new ReceivedStamp());
 
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
         $bus->expects($this->at(0))->method('dispatch')->with($envelope)->willReturn($envelope);
-        $retryStrategy = $this->getMockBuilder(RetryStrategyInterface::class)->getMock();
 
-        $worker = new Worker($receiver, $bus, 'receiver_id', $retryStrategy);
-        $worker->run();
+        $worker = new Worker([$receiver], $bus, []);
+        $worker->run([], function (?Envelope $envelope) use ($worker) {
+            // stop after the messages finish
+            if (null === $envelope) {
+                $worker->stop();
+            }
+        });
     }
 
     public function testDispatchCausesRetry()
     {
-        $receiver = new CallbackReceiver(function ($handler) {
-            $handler(new Envelope(new DummyMessage('Hello'), new SentStamp('Some\Sender', 'sender_alias')));
-        });
+        $receiver = new DummyReceiver([
+            [new Envelope(new DummyMessage('Hello'), new SentStamp('Some\Sender', 'sender_alias'))],
+        ]);
 
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
         $bus->expects($this->at(0))->method('dispatch')->willThrowException(new \InvalidArgumentException('Why not'));
@@ -93,8 +102,13 @@ class WorkerTest extends TestCase
         $retryStrategy = $this->getMockBuilder(RetryStrategyInterface::class)->getMock();
         $retryStrategy->expects($this->once())->method('isRetryable')->willReturn(true);
 
-        $worker = new Worker($receiver, $bus, 'receiver_id', $retryStrategy);
-        $worker->run();
+        $worker = new Worker(['receiver1' => $receiver], $bus, ['receiver1' => $retryStrategy]);
+        $worker->run([], function (?Envelope $envelope) use ($worker) {
+            // stop after the messages finish
+            if (null === $envelope) {
+                $worker->stop();
+            }
+        });
 
         // old message acknowledged
         $this->assertSame(1, $receiver->getAcknowledgeCount());
@@ -102,9 +116,9 @@ class WorkerTest extends TestCase
 
     public function testDispatchCausesRejectWhenNoRetry()
     {
-        $receiver = new CallbackReceiver(function ($handler) {
-            $handler(new Envelope(new DummyMessage('Hello'), new SentStamp('Some\Sender', 'sender_alias')));
-        });
+        $receiver = new DummyReceiver([
+            [new Envelope(new DummyMessage('Hello'), new SentStamp('Some\Sender', 'sender_alias'))],
+        ]);
 
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
         $bus->method('dispatch')->willThrowException(new \InvalidArgumentException('Why not'));
@@ -112,17 +126,22 @@ class WorkerTest extends TestCase
         $retryStrategy = $this->getMockBuilder(RetryStrategyInterface::class)->getMock();
         $retryStrategy->expects($this->once())->method('isRetryable')->willReturn(false);
 
-        $worker = new Worker($receiver, $bus, 'receiver_id', $retryStrategy);
-        $worker->run();
+        $worker = new Worker(['receiver1' => $receiver], $bus, ['receiver1' => $retryStrategy]);
+        $worker->run([], function (?Envelope $envelope) use ($worker) {
+            // stop after the messages finish
+            if (null === $envelope) {
+                $worker->stop();
+            }
+        });
         $this->assertSame(1, $receiver->getRejectCount());
         $this->assertSame(0, $receiver->getAcknowledgeCount());
     }
 
     public function testDispatchCausesRejectOnUnrecoverableMessage()
     {
-        $receiver = new CallbackReceiver(function ($handler) {
-            $handler(new Envelope(new DummyMessage('Hello')));
-        });
+        $receiver = new DummyReceiver([
+            [new Envelope(new DummyMessage('Hello'))],
+        ]);
 
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
         $bus->method('dispatch')->willThrowException(new UnrecoverableMessageHandlingException('Will never work'));
@@ -130,36 +149,42 @@ class WorkerTest extends TestCase
         $retryStrategy = $this->getMockBuilder(RetryStrategyInterface::class)->getMock();
         $retryStrategy->expects($this->never())->method('isRetryable');
 
-        $worker = new Worker($receiver, $bus, 'receiver_id', $retryStrategy);
-        $worker->run();
+        $worker = new Worker(['receiver1' => $receiver], $bus, ['receiver1' => $retryStrategy]);
+        $worker->run([], function (?Envelope $envelope) use ($worker) {
+            // stop after the messages finish
+            if (null === $envelope) {
+                $worker->stop();
+            }
+        });
         $this->assertSame(1, $receiver->getRejectCount());
     }
 
     public function testWorkerDoesNotSendNullMessagesToTheBus()
     {
-        $receiver = new CallbackReceiver(function ($handler) {
-            $handler(null);
-        });
+        $receiver = new DummyReceiver([
+            null,
+        ]);
 
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
         $bus->expects($this->never())->method('dispatch');
-        $retryStrategy = $this->getMockBuilder(RetryStrategyInterface::class)->getMock();
 
-        $worker = new Worker($receiver, $bus, 'receiver_id', $retryStrategy);
-        $worker->run();
+        $worker = new Worker([$receiver], $bus);
+        $worker->run([], function (?Envelope $envelope) use ($worker) {
+            // stop after the messages finish
+            if (null === $envelope) {
+                $worker->stop();
+            }
+        });
     }
 
     public function testWorkerDispatchesEventsOnSuccess()
     {
         $envelope = new Envelope(new DummyMessage('Hello'));
-        $receiver = new CallbackReceiver(function ($handler) use ($envelope) {
-            $handler($envelope);
-        });
+        $receiver = new DummyReceiver([[$envelope]]);
 
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
         $bus->method('dispatch')->willReturn($envelope);
 
-        $retryStrategy = $this->getMockBuilder(RetryStrategyInterface::class)->getMock();
         $eventDispatcher = $this->getMockBuilder(EventDispatcherInterface::class)->getMock();
 
         $eventDispatcher->expects($this->exactly(2))
@@ -169,22 +194,24 @@ class WorkerTest extends TestCase
                 [$this->isInstanceOf(WorkerMessageHandledEvent::class)]
             );
 
-        $worker = new Worker($receiver, $bus, 'receiver_id', $retryStrategy, $eventDispatcher);
-        $worker->run();
+        $worker = new Worker([$receiver], $bus, [], $eventDispatcher);
+        $worker->run([], function (?Envelope $envelope) use ($worker) {
+            // stop after the messages finish
+            if (null === $envelope) {
+                $worker->stop();
+            }
+        });
     }
 
     public function testWorkerDispatchesEventsOnError()
     {
         $envelope = new Envelope(new DummyMessage('Hello'));
-        $receiver = new CallbackReceiver(function ($handler) use ($envelope) {
-            $handler($envelope);
-        });
+        $receiver = new DummyReceiver([[$envelope]]);
 
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
         $exception = new \InvalidArgumentException('Oh no!');
         $bus->method('dispatch')->willThrowException($exception);
 
-        $retryStrategy = $this->getMockBuilder(RetryStrategyInterface::class)->getMock();
         $eventDispatcher = $this->getMockBuilder(EventDispatcherInterface::class)->getMock();
 
         $eventDispatcher->expects($this->exactly(2))
@@ -194,7 +221,143 @@ class WorkerTest extends TestCase
                 [$this->isInstanceOf(WorkerMessageFailedEvent::class)]
             );
 
-        $worker = new Worker($receiver, $bus, 'receiver_id', $retryStrategy, $eventDispatcher);
-        $worker->run();
+        $worker = new Worker([$receiver], $bus, [], $eventDispatcher);
+        $worker->run([], function (?Envelope $envelope) use ($worker) {
+            // stop after the messages finish
+            if (null === $envelope) {
+                $worker->stop();
+            }
+        });
+    }
+
+    public function testTimeoutIsConfigurable()
+    {
+        $apiMessage = new DummyMessage('API');
+        $receiver = new DummyReceiver([
+            [new Envelope($apiMessage), new Envelope($apiMessage)],
+            [], // will cause a wait
+            [], // will cause a wait
+            [new Envelope($apiMessage)],
+            [new Envelope($apiMessage)],
+            [], // will cause a wait
+            [new Envelope($apiMessage)],
+        ]);
+
+        $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
+
+        $worker = new Worker([$receiver], $bus);
+        $receivedCount = 0;
+        $startTime = microtime(true);
+        // sleep .1 after each idle
+        $worker->run(['sleep' => 100000], function (?Envelope $envelope) use ($worker, &$receivedCount, $startTime) {
+            if (null !== $envelope) {
+                ++$receivedCount;
+            }
+
+            if (5 === $receivedCount) {
+                $worker->stop();
+                $duration = microtime(true) - $startTime;
+
+                // wait time should be .3 seconds
+                // use .29 & .31 for timing "wiggle room"
+                $this->assertGreaterThanOrEqual(.29, $duration);
+                $this->assertLessThan(.31, $duration);
+            }
+        });
+    }
+
+    public function testWorkerWithMultipleReceivers()
+    {
+        // envelopes, in their expected delivery order
+        $envelope1 = new Envelope(new DummyMessage('message1'));
+        $envelope2 = new Envelope(new DummyMessage('message2'));
+        $envelope3 = new Envelope(new DummyMessage('message3'));
+        $envelope4 = new Envelope(new DummyMessage('message4'));
+        $envelope5 = new Envelope(new DummyMessage('message5'));
+        $envelope6 = new Envelope(new DummyMessage('message6'));
+
+        /*
+         * Round 1) receiver 1 & 2 have nothing, receiver 3 processes envelope1 and envelope2
+         * Round 2) receiver 1 has nothing, receiver 2 processes envelope3, receiver 3 is not called
+         * Round 3) receiver 1 processes envelope 4, receivers 2 & 3 are not called
+         * Round 4) receiver 1 processes envelope 5, receivers 2 & 3 are not called
+         * Round 5) receiver 1 has nothing, receiver 2 has nothing, receiver 3 has envelope 6
+         */
+        $receiver1 = new DummyReceiver([
+            [],
+            [],
+            [$envelope4],
+            [$envelope5],
+            [],
+        ]);
+        $receiver2 = new DummyReceiver([
+            [],
+            [$envelope3],
+            [],
+        ]);
+        $receiver3 = new DummyReceiver([
+            [$envelope1, $envelope2],
+            [],
+            [$envelope6],
+        ]);
+
+        $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
+
+        $receivedCount = 0;
+        $worker = new Worker([$receiver1, $receiver2, $receiver3], $bus);
+        $processedEnvelopes = [];
+        $worker->run([], function (?Envelope $envelope) use ($worker, &$receivedCount, &$processedEnvelopes) {
+            if (null !== $envelope) {
+                $processedEnvelopes[] = $envelope;
+                ++$receivedCount;
+            }
+
+            // stop after the messages finish
+            if (6 === $receivedCount) {
+                $worker->stop();
+            }
+        });
+
+        // make sure they were processed in the correct order
+        $this->assertSame([$envelope1, $envelope2, $envelope3, $envelope4, $envelope5, $envelope6], $processedEnvelopes);
+    }
+}
+
+class DummyReceiver implements ReceiverInterface
+{
+    private $deliveriesOfEnvelopes;
+    private $acknowledgeCount = 0;
+    private $rejectCount = 0;
+
+    public function __construct(array $deliveriesOfEnvelopes)
+    {
+        $this->deliveriesOfEnvelopes = $deliveriesOfEnvelopes;
+    }
+
+    public function get(): iterable
+    {
+        $val = array_shift($this->deliveriesOfEnvelopes);
+
+        return null === $val ? [] : $val;
+    }
+
+    public function ack(Envelope $envelope): void
+    {
+        ++$this->acknowledgeCount;
+    }
+
+    public function reject(Envelope $envelope): void
+    {
+        ++$this->rejectCount;
+    }
+
+    public function getAcknowledgeCount(): int
+    {
+        return $this->acknowledgeCount;
+    }
+
+    public function getRejectCount(): int
+    {
+        return $this->rejectCount;
     }
 }