+++ src/Symfony/Component/Messenger/Transport/Receiver/ReceiverInterface.php
@@ -25,21 +25,25 @@ interface ReceiverInterface
     /**
      * Receive some messages to the given handler.
      *
-     * The handler will have, as argument, the received {@link \Symfony\Component\Messenger\Envelope} containing the message.
-     * Note that this envelope can be `null` if the timeout to receive something has expired.
+     * While this method could return an unlimited number of messages,
+     * the intention is that it returns only one, or a "small number"
+     * of messages each time. This gives the user more flexibility:
+     * they can finish processing the one (or "small number") of messages
+     * from this receiver and move on to check other receivers for messages.
+     * If a this method returns too many messages, it could cause a
+     * blocking effect where handling the messages received from one
+     * call to get() takes a long time, blocking other receivers from
+     * being called.
      *
-     * If the received message cannot be decoded, the message should not
+     * If a received message cannot be decoded, the message should not
      * be retried again (e.g. if there's a queue, it should be removed)
      * and a MessageDecodingFailedException should be thrown.
      *
      * @throws TransportException If there is an issue communicating with the transport
+     *
+     * @return Envelope[]
      */
-    public function receive(callable $handler): void;
-
-    /**
-     * Stop receiving some messages.
-     */
-    public function stop(): void;
+    public function get(): iterable;
 
     /**
      * Acknowledge that the passed message was handled.